+++ spark/v3.1/spark/src/main/java/org/apache/iceberg/spark/actions/Spark3BinPackStrategy.java
@@ -65,6 +65,7 @@ public class Spark3BinPackStrategy extends BinPackStrategy {
           .option(SparkReadOptions.FILE_SCAN_TASK_SET_ID, groupID)
           .option(SparkReadOptions.SPLIT_SIZE, splitSize(inputFileSize(filesToRewrite)))
           .option(SparkReadOptions.FILE_OPEN_COST, "0")
+          .option(SparkReadOptions.PLAN_TASKS_IGNORE_DELETE_FILES_WEIGHT, "true")
           .load(table.name());
 
       // write the packed data into new files where each split becomes a new file
