+++ spark/v3.2/spark/src/main/java/org/apache/iceberg/spark/source/SparkFilesScan.java
@@ -40,6 +40,7 @@ class SparkFilesScan extends SparkBatchScan {
   private final long splitSize;
   private final int splitLookback;
   private final long splitOpenFileCost;
+  private final boolean planTasksIgnoreDeleteFilesWeight;
 
   private List<CombinedScanTask> tasks = null; // lazy cache of tasks
 
@@ -51,6 +52,8 @@ class SparkFilesScan extends SparkBatchScan {
     this.splitSize = readConf.splitSize();
     this.splitLookback = readConf.splitLookback();
     this.splitOpenFileCost = readConf.splitOpenFileCost();
+    this.planTasksIgnoreDeleteFilesWeight = options.getBoolean(
+        SparkReadOptions.PLAN_TASKS_IGNORE_DELETE_FILES_WEIGHT, false);
   }
 
   @Override
@@ -66,8 +69,7 @@ class SparkFilesScan extends SparkBatchScan {
           CloseableIterable.withNoopClose(files),
           splitSize);
       CloseableIterable<CombinedScanTask> scanTasks = TableScanUtil.planTasks(
-          splitFiles, splitSize,
-          splitLookback, splitOpenFileCost);
+          splitFiles, splitSize, splitLookback, splitOpenFileCost, planTasksIgnoreDeleteFilesWeight);
       this.tasks = Lists.newArrayList(scanTasks);
     }
